backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
trafilatura==1.12.2
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydub==0.25.1
openai>=1.30.0
python-dotenv==1.0.1
gunicorn==22.0.0
